% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremalmatching.R
\name{Amatrix}
\alias{Amatrix}
\title{Construct A matrix in dimension d}
\usage{
Amatrix(d, named = TRUE)
}
\arguments{
\item{d}{the dimension}

\item{named}{logical value to determine whether rows and columns are named}
}
\value{
a square matrix with 2^(d-1) rows and columns
}
\description{
Construct A matrix in dimension d
}
\examples{
Amatrix(4)
}
