% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attainable.R
\name{attainable}
\alias{attainable}
\title{Determine  mixture representation from partial concordance signature}
\usage{
attainable(sig, maximize = FALSE)
}
\arguments{
\item{sig}{partial concordance signature}

\item{maximize}{a logical variable stating whether unspecified concordance
probabilities should be jointly maximized rather than minimized which is the
default}
}
\value{
vector of weights of a possible extremal copula
}
\description{
Determine  mixture representation from partial concordance signature
}
\examples{
kappa2 <- 0.7
sig <- c(1, rep(kappa2, 6))
attainable(sig)
attainable(sig, maximize = TRUE)
}
