% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremalmatching.R
\name{Bmatrix}
\alias{Bmatrix}
\title{Construct B matrix in dimension d}
\usage{
Bmatrix(d, named = TRUE, scale = TRUE)
}
\arguments{
\item{d}{the dimension}

\item{named}{logical value to determine whether rows and columns are named}

\item{scale}{if this is set to false the final scaling of the columns by multiplicity
is not carried out}
}
\value{
a square matrix with 1 + floor(d/2) rows and columns
}
\description{
Construct B matrix in dimension d
}
\examples{
Bmatrix(7)
}
