% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremalcopula.R
\name{estsignature}
\alias{estsignature}
\title{Estimate a concordance signature}
\usage{
estsignature(data)
}
\arguments{
\item{data}{a matrix of data observations}
}
\value{
a vector giving the estimated concordance signature
}
\description{
Estimate a concordance signature
}
\examples{
data <- as.matrix(iris[,1:4])
estsignature(data)
}
