% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{evenpowerset}
\alias{evenpowerset}
\title{Lexicographically ordered even power set}
\usage{
evenpowerset(d)
}
\arguments{
\item{d}{the dimension of the set}
}
\value{
a character vector with the elements of the even power set
}
\description{
Lexicographically ordered even power set
}
\examples{
evenpowerset(4)
}
