% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremalcopula.R
\name{extcorr}
\alias{extcorr}
\title{Extremal correlation matrices}
\usage{
extcorr(k, d)
}
\arguments{
\item{k}{a vector of integers between one and 2^(d-1) specifying extremal
copulas in dimension d}

\item{d}{the dimension of the extremal correlation matrix}
}
\value{
a list of matrices if k is a vector or a single matrix if k is scalar
}
\description{
Extremal correlation matrices
}
\examples{
extcorr(2, 4)
extcorr(1:8, 4)
}
