% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{extmixcorr}
\alias{extmixcorr}
\title{Correlation matrix of extremal mixture copula}
\usage{
extmixcorr(w)
}
\arguments{
\item{w}{vector of weights specifying the extremal mixture copula}
}
\value{
correlation matrix of the extremal mixture copula
}
\description{
Correlation matrix of extremal mixture copula
}
\examples{
extmixcorr(rep(1, 8)/8)
}
