% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremalcopula.R
\name{extremaltable}
\alias{extremaltable}
\title{Make table of extremal representation}
\usage{
extremaltable(B)
}
\arguments{
\item{B}{full concordance signature}
}
\value{
a dataframe that can be readily turned into latex table or knitr table
}
\description{
Make table of extremal representation
}
\examples{
P <- 0.7 * matrix(1, nrow = 3, ncol = 3) + 0.3 * diag(3)
B <- consignature(P)
extremaltable(B)
}
