% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremalcopula.R
\name{findextremal}
\alias{findextremal}
\title{Determine extremal mixture representation from full concordance signature}
\usage{
findextremal(B, named = TRUE)
}
\arguments{
\item{B}{full concordance signature}

\item{named}{logical value to determine whether weights are named}
}
\value{
vector of weights of extremal copula
}
\description{
Determine extremal mixture representation from full concordance signature
}
\examples{
P <- 0.7 * matrix(1, nrow = 3, ncol = 3) + 0.3 * diag(3)
B <- consignature(P)
findextremal(B)
}
