% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremalcopula.R
\name{findpolytope}
\alias{findpolytope}
\title{Find polytope of attainable weights from partial concordance signature}
\usage{
findpolytope(sig, d)
}
\arguments{
\item{sig}{partial concordance signature}

\item{d}{dimension of copula}
}
\value{
a matrix where each row is a vertex of the polytope
}
\description{
Find polytope of attainable weights from partial concordance signature
}
\examples{
tau <- c(-0.19, -0.29, 0.49, -0.34, 0.3, -0.79)
kappa <- (1+tau)/2
findpolytope(c(1, kappa), d = 4)
}
