% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremalcopula.R
\name{Gaussorthant}
\alias{Gaussorthant}
\title{Orthant probability for a multivariate Gaussian vector}
\usage{
Gaussorthant(P)
}
\arguments{
\item{P}{correlation matrix of the normal distribution}
}
\value{
the lower orthant probability
}
\description{
Orthant probability for a multivariate Gaussian vector
}
\examples{
P <- 0.7 * matrix(1, nrow = 3, ncol = 3) + 0.3 * diag(3)
Gaussorthant(P)
}
