% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremalcopula.R
\name{quadrants}
\alias{quadrants}
\title{Identify quadrants of multivariate data points}
\usage{
quadrants(signs, diagonal = FALSE)
}
\arguments{
\item{signs}{a matrix containing signs of the data points}

\item{diagonal}{whether quadrants on same diagonal should be amalgamated}
}
\value{
a vector containing the quadrant of each point
}
\description{
Identify quadrants of multivariate data points
}
\examples{
data <- rmvnorm(n = 1000, sigma = diag(3))
table(quadrants(sign(data)))
table(quadrants(sign(data), diagonal = TRUE))
}
