% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{rextremal}
\alias{rextremal}
\title{Sample from extremal copulas}
\usage{
rextremal(k, d)
}
\arguments{
\item{k}{a vector of integers between one and 2^(d-1) specifying extremal
copulas in dimension d}

\item{d}{the dimension of the extremal copula}
}
\value{
a matrix with as many rows as the length of k
such that each row is a realisation from an extremal copula in dimension d
}
\description{
Sample from extremal copulas
}
\examples{
rextremal(rep(2, 10), 4)
}
