% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{rextremalmix}
\alias{rextremalmix}
\title{Sample from extremal mixture copula}
\usage{
rextremalmix(w, n = 1)
}
\arguments{
\item{w}{vector of weights specifing the extremal mixture copula}

\item{n}{number of realisations to generate}
}
\value{
a matrix with n rows where each row is a a realisation from the
extremal mixture copula
}
\description{
Sample from extremal mixture copula
}
\examples{
rextremalmix(rep(1, 8)/8, 100)
}
