% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{setpaste}
\alias{setpaste}
\title{Paste vector in set notation}
\usage{
setpaste(v)
}
\arguments{
\item{v}{a numeric vector}
}
\value{
a character string gving vector as set
}
\description{
Paste vector in set notation
}
