% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{signsofpairs}
\alias{signsofpairs}
\title{Compute signs of differences of pairs}
\usage{
signsofpairs(data)
}
\arguments{
\item{data}{a matrix of data of length n}
}
\value{
a matrix of length n*(n-1)/2
}
\description{
Compute signs of differences of pairs
}
\examples{
data <- mvtnorm::rmvnorm(n = 10, sigma = diag(3))
signsofpairs(data)
}
